chore:  some adjusts on new AST
