--- lx_erl_compile_server/README.md
@@ -1,9 +1,0 @@
-lx_erl_compile_server
-=====
-
-An OTP application
-
-Build
------
-
-    $ rebar3 compile
