--- lx_erl_compile_server/rebar.config
@@ -1,7 +1,0 @@
-{erl_opts, [debug_info]}.
-{deps, []}.
-
-{shell, [
-  % {config, "config/sys.config"},
-    {apps, [lx_erl_compile_server]}
-]}.
