--- lx_erl_compile_server/rebar.lock
@@ -1,1 +1,0 @@
-[].
