+++ src/tests/builtin_functions/hello_world.ex
@@ -1,6 +1,6 @@
 defmodule HelloWorld do
   def main() do
     Lx.IO.puts("Using builtin functions\n")
-    :ok
+    # :ok
   end
 end
