+++ src/tests/modules_test.v
@@ -11,6 +11,6 @@ pub fn test_modules_ex_file() {
 	mut bin := compiler_v.generate(path)
 
 	assert 'A' == bin.program.modules['A'].name
-	assert '{{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:B]},[{:do,{:def, [line: 2,type: SUM::float], [{:one, [line: 2,type: SUM::float], []},[{:do,:2.0}]]}}]]},{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:A]},[{:do,{:def, [line: 2,type: SUM::float], [{:main, [line: 2,type: SUM::float], []},[{:do,{{:., [line: 3,type: float], [{:__aliases__, [line: 3,type: float], [:B]},:one]}, [line: 3,type: float], []}}]]}}]]}}' == bin.program.modules['A'].str()
+	assert '{{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:B]},[{:do,{:def, [line: 2,type: SUM::float], [{:one, [line: 2,type: SUM::float], []},[{:do,2.0}]]}}]]},{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:A]},[{:do,{:def, [line: 2,type: SUM::float], [{:main, [line: 2,type: SUM::float], []},[{:do,{{:., [line: 3,type: float], [{:__aliases__, [line: 3,type: float], [:B]},:one]}, [line: 3,type: float], []}}]]}}]]}}' == bin.program.modules['A'].str()
 	assert '2.0' == compiler_v.execute(mut bin)
 }
