+++ src/tests/precedence_promote_test.v
@@ -10,6 +10,6 @@ pub fn test_precedence_promote_ex_file() {
 
 	// Generate HelloWorld Module
 	assert 'PrecedenceTest' == bin.program.modules['PrecedenceTest'].name
-	assert '{{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:PrecedenceTest]},[{:do,{:def, [line: 2,type: SUM::float], [{:main, [line: 2,type: SUM::float], []},[{:do,{:+, [line: 3,type: float], [:1.0,:2.0999999046325684]}}]]}}]]}}' == bin.program.modules['PrecedenceTest'].str()
-	assert '3.0999999046325684' == compiler_v.execute(mut bin)
+	assert '{{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:PrecedenceTest]},[{:do,{:def, [line: 2,type: SUM::float], [{:main, [line: 2,type: SUM::float], []},[{:do,{:+, [line: 3,type: float], [1.0,2.1]}}]]}}]]}}' == bin.program.modules['PrecedenceTest'].str()
+	assert '3.1' == compiler_v.execute(mut bin)
 }
