+++ src/tests/precedence_test.v
@@ -10,6 +10,6 @@ pub fn test_precedence_ex_file() {
 
 	// Generate HelloWorld Module
 	assert 'PrecedenceTest' == bin.program.modules['PrecedenceTest'].name
-	assert '{{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:PrecedenceTest]},[{:do,{:def, [line: 2,type: SUM::integer], [{:main, [line: 2,type: SUM::integer], []},[{:do,{:+, [line: 3,type: integer], [{:-, [line: 3,type: integer], [{:-, [line: 3,type: integer], [{:+, [line: 3,type: integer], [{:+, [line: 3,type: integer], [:1,:2]},:3]},:2]},{:/, [line: 3,type: integer], [{:*, [line: 3,type: integer], [{:/, [line: 3,type: integer], [:5,:2]},:3]},:4]}]},{:/, [line: 3,type: integer], [:5,{:+, [line: 3,inside_parens: true,type: integer], [:1,:2]}]}]}}]]}}]]}}' == bin.program.modules['PrecedenceTest'].str()
+	assert '{{:defmodule, [line: 1], [{:__aliases__, [line: 1], [:PrecedenceTest]},[{:do,{:def, [line: 2,type: SUM::integer], [{:main, [line: 2,type: SUM::integer], []},[{:do,{:+, [line: 3,type: integer], [{:-, [line: 3,type: integer], [{:-, [line: 3,type: integer], [{:+, [line: 3,type: integer], [{:+, [line: 3,type: integer], [1,2]},3]},2]},{:/, [line: 3,type: integer], [{:*, [line: 3,type: integer], [{:/, [line: 3,type: integer], [5,2]},3]},4]}]},{:/, [line: 3,type: integer], [5,{:+, [line: 3,inside_parens: true,type: integer], [1,2]}]}]}}]]}}]]}}' == bin.program.modules['PrecedenceTest'].str()
 	assert '4' == compiler_v.execute(mut bin)
 }
