+++ src/tests/strings_test/strings.ex
@@ -1,0 +1,31 @@
+defmodule StringsTest do
+
+  def main() do
+
+    # ## String literal
+    # "Hello Lx"
+    # "Hello
+    # Lx"
+    # "👩‍💻 こんにちは Lx 💫"
+
+    # ## MultilineString literal
+    # """
+    #   Multiline String
+    #   Other Line
+    #   And other line
+    # """
+
+    # ### Concatenation
+    # "name:" <> "Fulano"
+    # name = "Fulano"
+    # "Hello, " <> name <> "!"
+
+    # ### Interpolation
+    # "Hello, #{name}"
+
+    # ### Scape Sequences
+    # # "Here is a double quote -> \" <-"
+    # # "C:\\Users\\Lx"
+    :ok
+  end
+end
